Special case Bundler earlier
Some changes in Bundler 2.3.20 [1] have created a little inconsistency,
because `Bundler::Definition#resolve` will now inconsistently include
the Bundler gem itself. If the resolve is taken exclusively from the
lockfile, then it's not included, because `Gemfile.lock` files don't
lock the Bundler gem itself even if it's a dependency (only in the
BUNDLED WITH section). However, when resolving from scratch, then
`Bundler::Definition#resolve` will include Bundler if Bundler is a
Gemfile dependency.

Dependabot was making the assumption that the Bundler gem itself is
always included in `Bundler::Definition#resolve`, and thus the outcome
of one spec now changes, because the way we short-circuit logic for
Bundler no longer works.

I'm not a fan of the inconsistency created on the Bundler side (the
mentioned commit was meant to reduce inconsistencies, not the opposite).
However, this can be easily fixed in dependabot itself by
short-circuiting earlier in the Bundler case, before even looking at the
definition.

[1] https://github.com/rubygems/rubygems/pull/5784/commits/2777e79b8efe1bf77589496632c74464022f04f3